requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
boto3==1.34.35
botocore==1.34.35
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.1
Deprecated==1.2.14
deprecation==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
img2pdf==0.5.1
jmespath==1.0.1
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
ocrmypdf==16.0.4
packaging==23.2
pdf2image==1.17.0
pdfminer.six==20231228
pikepdf==8.11.2
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyMuPDF==1.23.20
PyMuPDFb==1.23.9
pypng==0.20220715.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
qrcode==7.4.2
reportlab==4.0.9
rich==13.7.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
wrapt==1.16.0
